Verificar por atualização de dependências ao iniciar bot (repl.it)

+++ deploy_handlers/replit.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+[ -z "$VIDEO_PREVIEW" ] || { . venv/bin/activate && python3 preview.py; kill "$PPID"; exit 1; }
+
 if [ ! -d "venv" ] || [ ! -f "./venv/.deployed" ]; then
   bash quick_update.sh
   rm -rf venv
@@ -8,13 +10,10 @@ if [ ! -d "venv" ] || [ ! -f "./venv/.deployed" ]; then
   . venv/bin/activate
   python3 -m pip config unset --user install.use-feature
   python3 -m pip install -U pip poetry
-  poetry install
   touch ./venv/.deployed
 else
-  python3 -m venv venv # tempfix
   . venv/bin/activate
 fi
 
-[ -z "$VIDEO_PREVIEW" ] || { python3 preview.py; kill "$PPID"; exit 1; }
-
+poetry install
 python3 main.py
